Reorganize internals into a Protocol namespace.
