+++ lib/Protocol/Message.php
@@ -1,5 +1,7 @@
 <?php
-namespace ParagonIE\Gossamer;
+namespace ParagonIE\Gossamer\Protocol;
+
+use ParagonIE\Gossamer\Util;
 
 /**
  * Class Message
